Milestone #2 - Custom hook addition for pomdoro timer
