+++ .gitignore
@@ -64,3 +64,5 @@ yarn-error.log
 
 # testing
 /coverage
+
+.env
