+++ src/components/timer/index.tsx
@@ -1,9 +1,9 @@
 /* eslint-disable prettier/prettier */
-/* eslint-disable react-hooks/exhaustive-deps */
-import React, {useEffect, useState, useRef} from 'react';
+import React from 'react';
 import {Text, TouchableOpacity, View} from 'react-native';
 import {formatTime} from '../../utils/formatTime';
 import {timerStyles} from './style';
+import usePomodoro from '../../hooks/usePomodoro';
 
 interface TimerProps {
   workDuration: number;
@@ -11,59 +11,21 @@ interface TimerProps {
 }
 
 const Timer: React.FC<TimerProps> = ({workDuration, breakDuration}) => {
-  const [remainingTime, setRemainingTime] = useState(workDuration);
-  const [isWorking, setIsWorking] = useState(true);
-  const [isRunning, setIsRunning] = useState(false);
-  const intervalRef = useRef<NodeJS.Timeout | null>(null);
-
-  useEffect(() => {
-    const handleTick = () => {
-      setRemainingTime(prevTime => {
-        if (prevTime > 0) {
-          return prevTime - 1;
-        } else {
-          setIsWorking(prevIsWorking => {
-            if (prevIsWorking) {
-              setIsWorking(false);
-              setRemainingTime(breakDuration);
-            } else {
-              setIsWorking(true);
-              setRemainingTime(workDuration);
-            }
-            return prevIsWorking;
-          });
-          return prevTime;
-        }
-      });
-    };
-
-    if (isRunning) {
-      intervalRef.current = setInterval(handleTick, 1000);
-    }
-
-    return () => {
-      if (intervalRef.current) {
-        clearInterval(intervalRef.current);
-        intervalRef.current = null;
-      }
-    };
-  }, [isRunning, workDuration, breakDuration]);
-
-  useEffect(() => {
-    if (!isRunning) {
-      setIsWorking(true);
-      setRemainingTime(workDuration);
-    }
-  }, [workDuration, breakDuration]);
+  const {
+    remainingTime,
+    isWorking,
+    isRunning,
+    startTimer,
+    stopTimer,
+    resetTimer,
+  } = usePomodoro(workDuration, breakDuration);
 
   const timerToggle = () => {
-    setIsRunning(!isRunning);
-  };
-
-  const timerReset = () => {
-    setIsRunning(false);
-    setIsWorking(true);
-    setRemainingTime(workDuration);
+    if (isRunning) {
+      stopTimer();
+    } else {
+      startTimer();
+    }
   };
 
   return (
@@ -80,7 +42,7 @@ const Timer: React.FC<TimerProps> = ({workDuration, breakDuration}) => {
           </Text>
         </View>
       </TouchableOpacity>
-      <TouchableOpacity onPress={timerReset}>
+      <TouchableOpacity onPress={resetTimer}>
         <View style={timerStyles.resetButton}>
           <Text style={timerStyles.resetButtonText}>Reset</Text>
         </View>
